[#2246] Move Users.jsx|scss to users folder
